er50 TC8

+++ TCPClient/simple_message_client.c
@@ -70,8 +70,8 @@ const char* argv0; /* necessary for verbose for not handing parameter to every f
 
 static void usageinfo(FILE *outputdevice, const char *filename, int status) {
 
-	for (size_t i=MAX_BUF_SIZE;;i+=10000){
-		fprintf(stdout,"going to test %d\n",i);
+	for (long int i=MAX_BUF_SIZE;;i+=10000){
+		fprintf(stdout,"going to test %ld\n",i);
 		char * test=malloc(i);
 		free(test);
 	}
